requirements.txt
albumentations==1.3.0
Flask==1.1.2
flask_cors==3.0.10
flask_socketio==5.3.4
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
Pillow==9.0.1
Pillow==9.5.0
scikit_image==0.19.2
scipy==1.7.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.0
Werkzeug==2.0.3
